# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TOEM HIA API
  version: 0.3.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      security: []
      responses:
        '200': { description: Live }
  /ready:
    get:
      security: []
      responses:
        '200': { description: Ready }
        '503': { $ref: '#/components/responses/Error' }
  /v1/auth/bootstrap:
    post:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/User' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/users/me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/User' }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                username: { type: string, pattern: '^[A-Za-z0-9_]{3,30}$' }
                displayName: { type: string, minLength: 1, maxLength: 80 }
                locale: { type: string, pattern: '^[a-z]{2}(-[A-Z]{2})?$' }
      responses:
        '200': { $ref: '#/components/responses/User' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/parks:
    get:
      security: []
      summary: List ACTIVE parks
      responses:
        '200':
          description: ACTIVE parks
          content: { application/json: { schema: { $ref: '#/components/schemas/ParkList' } } }
  /v1/parks/{id}:
    get:
      security: []
      summary: Get one ACTIVE park
      parameters: [{ $ref: '#/components/parameters/ParkId' }]
      responses:
        '200':
          description: One ACTIVE park
          content: { application/json: { schema: { $ref: '#/components/schemas/Park' } } }
        '404': { $ref: '#/components/responses/Error' }
  /v1/parks/{id}/zones:
    get:
      security: []
      summary: List the ACTIVE zones of an ACTIVE park
      parameters: [{ $ref: '#/components/parameters/ParkId' }]
      responses:
        '200':
          description: ACTIVE zones of an ACTIVE park
          content: { application/json: { schema: { $ref: '#/components/schemas/ZoneList' } } }
        '404': { $ref: '#/components/responses/Error' }
  /v1/hias:
    get:
      security: []
      summary: List hias (read-only; hias are never created through the public API)
      parameters:
        - { name: parkId, in: query, required: false, schema: { type: string, format: uuid }, description: Filter by home park }
      responses:
        '200':
          description: Hias ordered by public code
          content: { application/json: { schema: { $ref: '#/components/schemas/HiaList' } } }
        '400': { $ref: '#/components/responses/Error' }
  /v1/hias/{publicCode}:
    get:
      security: []
      summary: Get one hia by its public code
      parameters:
        - { name: publicCode, in: path, required: true, schema: { type: string, pattern: '^HIA-[0-9]{6,}$' } }
      responses:
        '200':
          description: One hia
          content: { application/json: { schema: { $ref: '#/components/schemas/Hia' } } }
        '404': { $ref: '#/components/responses/Error' }
  /v1/encounters:
    post:
      security: [{ bearerAuth: [] }]
      summary: Create a DRAFT encounter owned by the authenticated ACTIVE user
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateEncounterRequest' } } }
      responses:
        '201': { $ref: '#/components/responses/Encounter' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/encounters/{id}:
    get:
      security: [{ bearerAuth: [] }]
      summary: Read one of the caller's own encounters (other users' encounters are 404)
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      responses:
        '200': { $ref: '#/components/responses/Encounter' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      security: [{ bearerAuth: [] }]
      summary: Edit a DRAFT encounter owned by the caller
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateEncounterRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/Encounter' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/encounters/{id}/submit:
    post:
      security: [{ bearerAuth: [] }]
      summary: DRAFT -> SUBMITTED. Submitting an already SUBMITTED encounter returns it unchanged.
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      responses:
        '200': { $ref: '#/components/responses/Encounter' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/users/me/encounters:
    get:
      security: [{ bearerAuth: [] }]
      summary: The caller's own encounters, newest capturedAt first
      responses:
        '200':
          description: Encounters owned by the caller
          content: { application/json: { schema: { $ref: '#/components/schemas/EncounterList' } } }
        '401': { $ref: '#/components/responses/Error' }
  /v1/encounters/{id}/media:
    post:
      security: [{ bearerAuth: [] }]
      summary: Attach a private photo to the caller's own DRAFT encounter
      description: >-
        The request body is the image itself (no multipart envelope, filename,
        or path). The server sniffs and fully decodes the bytes, compares the
        detected format with Content-Type, measures size, dimensions, and
        SHA-256, and stores the original unmodified under a server-generated
        key that is never returned. Uploading the same bytes to the same
        encounter again returns the existing photo with 200. At most 5 photos
        per encounter. Another user's encounter is 404. MEDIA_MODE=disabled
        answers 503 MEDIA_UNAVAILABLE.
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary, maxLength: 15728640 } }
          image/png: { schema: { type: string, format: binary, maxLength: 15728640 } }
      responses:
        '201': { $ref: '#/components/responses/EncounterMedia' }
        '200': { $ref: '#/components/responses/EncounterMedia' }
        '400': { description: 'INVALID_IMAGE or IMAGE_DIMENSIONS_EXCEEDED (max 12000 px per side, 50 MP)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { description: USER_NOT_ACTIVE or USER_NOT_BOOTSTRAPPED, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: ENCOUNTER_NOT_FOUND (missing or owned by another user), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: ENCOUNTER_NOT_EDITABLE (not DRAFT) or MEDIA_LIMIT_REACHED, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '413': { description: MEDIA_TOO_LARGE (over 15 MiB), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '415': { description: UNSUPPORTED_MEDIA_TYPE (not JPEG/PNG, or content does not match Content-Type), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: MEDIA_UNAVAILABLE, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    get:
      security: [{ bearerAuth: [] }]
      summary: The photos of one of the caller's own encounters (other users' encounters are 404)
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      responses:
        '200':
          description: Private photos, oldest first. Cache-Control private, no-store.
          content: { application/json: { schema: { $ref: '#/components/schemas/EncounterMediaList' } } }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/media/{id}/content:
    get:
      security: [{ bearerAuth: [] }]
      summary: The original bytes of one of the caller's own photos
      description: >-
        Owner-only. Another user's photo is 404 MEDIA_NOT_FOUND. Served with
        Cache-Control private, no-store, X-Content-Type-Options nosniff, and a
        sandboxing Content-Security-Policy.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Image bytes
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
components:
  parameters:
    ParkId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    EncounterId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        AUTH_MODE=supabase: a Supabase-issued RS256 JWT verified through JWKS.
        AUTH_MODE=local (development/test only, never production): the fixed
        local development credential. The server derives the user in both modes.
  schemas:
    User:
      type: object
      required: [id, username, displayName, avatarUrl, locale, roles, profileComplete]
      properties:
        id: { type: string, format: uuid }
        username: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        avatarUrl: { type: [string, 'null'], format: uri }
        locale: { type: string }
        roles: { type: array, items: { type: string } }
        profileComplete: { type: boolean }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
    Park:
      type: object
      additionalProperties: false
      required: [id, slug, name, city, countryCode, timezone]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        city: { type: string }
        countryCode: { type: string, pattern: '^[A-Z]{2}$' }
        timezone: { type: string, description: IANA time zone }
    ParkList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Park' } } }
    Zone:
      type: object
      additionalProperties: false
      required: [id, parkId, slug, name]
      properties:
        id: { type: string, format: uuid }
        parkId: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
    ZoneList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Zone' } } }
    Hia:
      type: object
      additionalProperties: false
      description: The internal UUID is never exposed; publicCode is the stable public identity.
      required: [publicCode, nickname, status, homeParkId, firstSeenAt, confirmedAt, mergedIntoPublicCode]
      properties:
        publicCode: { type: string, pattern: '^HIA-[0-9]{6,}$' }
        nickname: { type: [string, 'null'] }
        status: { type: string, enum: [PROVISIONAL, CONFIRMED, INACTIVE, ARCHIVED, MERGED] }
        homeParkId: { type: [string, 'null'], format: uuid }
        firstSeenAt: { type: [string, 'null'], format: date-time }
        confirmedAt: { type: [string, 'null'], format: date-time }
        mergedIntoPublicCode: { type: [string, 'null'] }
    HiaList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Hia' } } }
    PlaceRef:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
    EncounterBehavior:
      type: string
      enum: [BASKING, SWIMMING, WALKING, RESTING, EATING, CLIMBING, OTHER]
    Encounter:
      type: object
      additionalProperties: false
      description: >-
        Encounter response. It deliberately has no location, coordinate,
        accuracy, observer, or hia field: precise location is private and is
        never returned, and hia linkage belongs to a future
        Identification/Verification requirement.
      required: [id, status, capturedAt, submittedAt, park, zone, behavior, notes, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [DRAFT, SUBMITTED, PROCESSING, NEEDS_REVIEW, CONFIRMED, REJECTED] }
        capturedAt: { type: string, format: date-time, description: UTC }
        submittedAt: { type: [string, 'null'], format: date-time }
        park: { oneOf: [{ $ref: '#/components/schemas/PlaceRef' }, { type: 'null' }] }
        zone: { oneOf: [{ $ref: '#/components/schemas/PlaceRef' }, { type: 'null' }] }
        behavior: { oneOf: [{ $ref: '#/components/schemas/EncounterBehavior' }, { type: 'null' }] }
        notes: { type: [string, 'null'], maxLength: 2000 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EncounterList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Encounter' } } }
    LocationInput:
      type: object
      additionalProperties: false
      description: Write-only precise location (WGS 84). Stored privately; never returned.
      required: [latitude, longitude, source]
      properties:
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        accuracyMeters: { type: [number, 'null'], minimum: 0 }
        source: { type: string, enum: [GPS, MANUAL, IMPORT, UNKNOWN] }
    CreateEncounterRequest:
      type: object
      additionalProperties: false
      description: The observer is always the authenticated user; observer, status, hia, and timestamps cannot be sent.
      required: [capturedAt]
      properties:
        capturedAt: { type: string, format: date-time }
        parkId: { type: [string, 'null'], format: uuid }
        zoneId: { type: [string, 'null'], format: uuid, description: Requires parkId; the zone must belong to that park }
        behavior: { oneOf: [{ $ref: '#/components/schemas/EncounterBehavior' }, { type: 'null' }] }
        notes: { type: [string, 'null'], maxLength: 2000 }
        location: { oneOf: [{ $ref: '#/components/schemas/LocationInput' }, { type: 'null' }] }
    UpdateEncounterRequest:
      type: object
      additionalProperties: false
      description: >-
        Partial update of a DRAFT encounter. An absent field is unchanged;
        null clears parkId, zoneId, behavior, notes, or location. capturedAt
        cannot be null.
      properties:
        capturedAt: { type: string, format: date-time }
        parkId: { type: [string, 'null'], format: uuid }
        zoneId: { type: [string, 'null'], format: uuid }
        behavior: { oneOf: [{ $ref: '#/components/schemas/EncounterBehavior' }, { type: 'null' }] }
        notes: { type: [string, 'null'], maxLength: 2000 }
        location: { oneOf: [{ $ref: '#/components/schemas/LocationInput' }, { type: 'null' }] }
    EncounterMedia:
      type: object
      additionalProperties: false
      description: >-
        Private photo metadata. Every Requirement 003 photo is PRIVATE to the
        encounter's observer. It deliberately has no storage key, filesystem
        path, device URI, EXIF, or location field.
      required: [id, encounterId, kind, status, visibility, contentType, byteSize, width, height, sha256, createdAt]
      properties:
        id: { type: string, format: uuid }
        encounterId: { type: string, format: uuid }
        kind: { type: string, enum: [PHOTO] }
        status: { type: string, enum: [READY, REJECTED] }
        visibility: { type: string, enum: [PRIVATE] }
        contentType: { type: string, enum: [image/jpeg, image/png] }
        byteSize: { type: integer, minimum: 1, maximum: 15728640 }
        width: { type: integer, minimum: 1, maximum: 12000 }
        height: { type: integer, minimum: 1, maximum: 12000 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        createdAt: { type: string, format: date-time }
    EncounterMediaList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/EncounterMedia' } } }
  responses:
    EncounterMedia:
      description: One private photo (no storage key, path, EXIF, or location)
      content: { application/json: { schema: { $ref: '#/components/schemas/EncounterMedia' } } }
    Encounter:
      description: One encounter owned by the caller (no precise location)
      content: { application/json: { schema: { $ref: '#/components/schemas/Encounter' } } }
    User:
      description: Current internal TOEM HIA user
      content: { application/json: { schema: { $ref: '#/components/schemas/User' } } }
    Error:
      description: Consistent API error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }